refactor(nika-http): the test code gets its own file — 1319 → 778
The next file on the cohesion ladder after #380. First draft split the
e2e mod to src/e2e.rs — and the unwrap/expect vector flagged its
test-only expects as PRODUCTION, because the prod-LOC convention the
hygiene vectors share exempts exactly ONE basename: tests.rs. The
convention working as designed; the split now follows it: src/tests.rs
= the unit invariants + the loopback e2e as a submodule, lib.rs keeps
one declaration line.

49/49 lib tests · unwrap+expect vectors green · clippy -D clean ·
public API byte-identical against the committed ubuntu baseline (a
local toolchain renders the tower-http Sized clause differently — the
baseline stays canonical, untouched).
