+++ crates/nika-http/src/lib.rs
@@ -831,590 +831,4 @@ impl Stream for CappedStream {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    // Public-type assertions live in `tests/http_contract.rs` — here
-    // only crate-private invariants.
-
-    #[test]
-    fn config_default_is_enforce() {
-        assert_eq!(HttpConfig::default().ssrf, SsrfMode::Enforce);
-    }
-
-    #[test]
-    fn net_default_is_unbounded() {
-        // No declared boundary by default — today's behavior (the SSRF
-        // floor is the only net guard until a `permits.net.http` is wired).
-        assert_eq!(HttpConfig::default().net, NetBoundary::Unbounded);
-    }
-
-    #[test]
-    fn net_boundary_gates_on_the_parsed_connect_host() {
-        // The RUNTIME half of permits.net.http (NIKA-SEC-004) — it gates on
-        // the PARSED url::Url host (what reqwest connects to), so a string-
-        // parser-vs-url-crate disagreement cannot bypass it.
-        let allow =
-            NetBoundary::Declared(vec!["allowed.com".to_owned(), "*.github.com".to_owned()]);
-        let u = |s: &str| url::Url::parse(s).expect("valid url");
-        // in-bounds (exact + `*.` subdomain)
-        assert!(check_net_allowlist(&allow, &u("https://allowed.com/p")).is_ok());
-        assert!(check_net_allowlist(&allow, &u("https://api.github.com/p")).is_ok());
-        // out-of-bounds → denied
-        assert!(matches!(
-            check_net_allowlist(&allow, &u("https://evil.com/p")),
-            Err(HttpError::HostNotAllowed { host }) if host == "evil.com"
-        ));
-        // THE BYPASS (rust-security P0): `\@` — a string parser reads host
-        // "allowed.com", but the connect host is "evil.com" (the `\` is a
-        // WHATWG path separator for http/https). Gating on url::Url denies it.
-        assert!(
-            matches!(
-                check_net_allowlist(&allow, &u(r"https://evil.com\@allowed.com/p")),
-                Err(HttpError::HostNotAllowed { host }) if host == "evil.com"
-            ),
-            "the backslash-userinfo confusion must resolve to the real connect host"
-        );
-        // userinfo `@` — the real host is evil.com (the classic confusion).
-        assert!(check_net_allowlist(&allow, &u("https://allowed.com@evil.com/p")).is_err());
-        // case-insensitive (the url crate lowercases) — no false block.
-        assert!(check_net_allowlist(&allow, &u("https://ALLOWED.COM/p")).is_ok());
-        // IPv6 matched bracket-free (permits write `::1`, never `[::1]`).
-        let v6 = NetBoundary::Declared(vec!["::1".to_owned()]);
-        assert!(check_net_allowlist(&v6, &u("http://[::1]:8080/p")).is_ok());
-        // trailing-dot FQDN (`allowed.com.`) ≡ `allowed.com` — no false block.
-        assert!(check_net_allowlist(&allow, &u("https://allowed.com./p")).is_ok());
-        // Unbounded = no declared boundary → the SSRF floor guards.
-        assert!(
-            check_net_allowlist(&NetBoundary::Unbounded, &u("https://anywhere.example/p")).is_ok()
-        );
-        // An empty Declared list (a `permits:` block that omits `net`) denies all.
-        assert!(
-            check_net_allowlist(
-                &NetBoundary::Declared(Vec::new()),
-                &u("https://allowed.com/p")
-            )
-            .is_err()
-        );
-    }
-
-    #[test]
-    fn host_of_matches_the_shared_parity_vectors() {
-        // The runtime extractor (`host_of`) MUST agree with the static
-        // checker (`nika-schema`'s `url_host`) on every shared vector — the
-        // no-drift guarantee. nika-schema asserts the SAME table against its
-        // extractor; if either drifts (`\@`, case, IPv6, trailing dot), one
-        // of the two suites fails. Single source of cases · zero drift.
-        for (input, expected) in nika_types::net::HOST_EXTRACTION_VECTORS {
-            let got = url::Url::parse(input).ok().and_then(|u| host_of(&u));
-            assert_eq!(got.as_deref(), *expected, "host_of disagrees on {input}");
-        }
-    }
-
-    #[test]
-    fn repeated_headers_comma_join_except_set_cookie() {
-        // RFC 9110 §5.2: N field lines ≡ one comma-joined value — the
-        // i18n shape (one Link per locale) must survive the lowering.
-        let mut headers = reqwest::header::HeaderMap::new();
-        headers.append(
-            "link",
-            "<https://x.test/fr>; rel=\"alternate\"; hreflang=\"fr\""
-                .parse()
-                .expect("valid"),
-        );
-        headers.append(
-            "link",
-            "<https://x.test/de>; rel=\"alternate\"; hreflang=\"de\""
-                .parse()
-                .expect("valid"),
-        );
-        // set-cookie: comma-UNSAFE (Expires dates) → last wins.
-        headers.append(
-            "set-cookie",
-            "a=1; Expires=Wed, 01 Jan 2025".parse().expect("valid"),
-        );
-        headers.append("set-cookie", "b=2".parse().expect("valid"));
-        // Non-UTF-8 values are skipped, not joined.
-        headers.append(
-            "x-binary",
-            reqwest::header::HeaderValue::from_bytes(&[0xff, 0xfe]).expect("bytes"),
-        );
-
-        let map = headers_to_btreemap(&headers);
-        assert_eq!(
-            map.get("link").map(String::as_str),
-            Some(
-                "<https://x.test/fr>; rel=\"alternate\"; hreflang=\"fr\", \
-                 <https://x.test/de>; rel=\"alternate\"; hreflang=\"de\""
-            ),
-            "both locales survive, comma-joined"
-        );
-        assert_eq!(map.get("set-cookie").map(String::as_str), Some("b=2"));
-        assert!(!map.contains_key("x-binary"), "non-UTF-8 skipped");
-    }
-
-    #[test]
-    fn method_mapping_covers_all_current_variants() {
-        for (kernel, expect) in [
-            (HttpMethod::Get, reqwest::Method::GET),
-            (HttpMethod::Post, reqwest::Method::POST),
-            (HttpMethod::Put, reqwest::Method::PUT),
-            (HttpMethod::Patch, reqwest::Method::PATCH),
-            (HttpMethod::Delete, reqwest::Method::DELETE),
-            (HttpMethod::Head, reqwest::Method::HEAD),
-            (HttpMethod::Options, reqwest::Method::OPTIONS),
-        ] {
-            assert_eq!(to_reqwest_method(kernel).expect("supported"), expect);
-        }
-    }
-
-    #[test]
-    fn timeout_mapping_saturates_huge_durations() {
-        let err = map_send_error_probe(Duration::from_secs(u64::MAX / 1000));
-        assert!(
-            matches!(err, HttpError::Timeout { duration_ms } if duration_ms > 0),
-            "expected a positive Timeout, got {err:?}"
-        );
-    }
-
-    #[test]
-    fn classify_resolved_flags_any_private_address() {
-        use std::net::{IpAddr, Ipv4Addr, SocketAddr};
-        let mixed = vec![
-            SocketAddr::new(IpAddr::V4(Ipv4Addr::new(8, 8, 8, 8)), 80),
-            SocketAddr::new(IpAddr::V4(Ipv4Addr::new(10, 0, 0, 1)), 80),
-        ];
-        assert!(matches!(classify_resolved(mixed), ResolveVerdict::Private));
-    }
-
-    #[test]
-    fn classify_resolved_all_public_passes() {
-        use std::net::{IpAddr, Ipv4Addr, SocketAddr};
-        let public = vec![
-            SocketAddr::new(IpAddr::V4(Ipv4Addr::new(8, 8, 8, 8)), 443),
-            SocketAddr::new(IpAddr::V4(Ipv4Addr::new(1, 1, 1, 1)), 443),
-        ];
-        assert!(matches!(
-            classify_resolved(public),
-            ResolveVerdict::AllPublic
-        ));
-    }
-
-    #[test]
-    fn classify_resolved_empty_is_empty() {
-        assert!(matches!(
-            classify_resolved(std::iter::empty()),
-            ResolveVerdict::Empty
-        ));
-    }
-
-    #[test]
-    fn classify_resolved_blocks_rebind_to_metadata_ip() {
-        use std::net::{IpAddr, Ipv4Addr, SocketAddr};
-        // The cloud-metadata link-local address a DNS-rebind would return.
-        let rebind = vec![SocketAddr::new(
-            IpAddr::V4(Ipv4Addr::new(169, 254, 169, 254)),
-            80,
-        )];
-        assert!(matches!(classify_resolved(rebind), ResolveVerdict::Private));
-    }
-
-    #[test]
-    fn strip_sensitive_removes_credentials_case_insensitive() {
-        let mut h = std::collections::BTreeMap::new();
-        h.insert("Authorization".to_string(), "Bearer x".to_string());
-        h.insert("COOKIE".to_string(), "s=1".to_string());
-        h.insert("X-Keep".to_string(), "ok".to_string());
-        strip_sensitive_headers(&mut h);
-        assert_eq!(
-            h.keys().map(String::as_str).collect::<Vec<_>>(),
-            vec!["X-Keep"]
-        );
-    }
-
-    #[test]
-    fn strip_body_removes_content_descriptors() {
-        let mut h = std::collections::BTreeMap::new();
-        h.insert("Content-Type".to_string(), "application/json".to_string());
-        h.insert("X-Keep".to_string(), "ok".to_string());
-        strip_body_headers(&mut h);
-        assert_eq!(
-            h.keys().map(String::as_str).collect::<Vec<_>>(),
-            vec!["X-Keep"]
-        );
-    }
-
-    /// Build a timeout-flavoured mapping without a real `reqwest::Error`
-    /// (reqwest errors cannot be constructed directly): exercise the
-    /// saturation arithmetic the mapper relies on.
-    fn map_send_error_probe(timeout: Duration) -> HttpError {
-        HttpError::Timeout {
-            duration_ms: u64::try_from(timeout.as_millis()).unwrap_or(u64::MAX),
-        }
-    }
-
-    /// A two-level wrapper replicating reqwest's connect-error nesting
-    /// (outer error → middle wrapper → the resolver's `HttpError`).
-    #[derive(Debug)]
-    struct Wrap {
-        msg: &'static str,
-        source: Box<dyn std::error::Error + Send + Sync>, // box-dyn-ok(test-harness): #[cfg(test)] fixture replicating reqwest's nesting
-    }
-    impl std::fmt::Display for Wrap {
-        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            f.write_str(self.msg)
-        }
-    }
-    impl std::error::Error for Wrap {
-        fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-            Some(self.source.as_ref())
-        }
-    }
-
-    #[test]
-    fn find_http_error_digs_a_nested_chain() {
-        let chain = Wrap {
-            msg: "client error (Connect)",
-            source: Box::new(Wrap {
-                msg: "dns error",
-                source: Box::new(HttpError::SsrfBlocked {
-                    url: "internal.corp".to_owned(),
-                }),
-            }),
-        };
-        let found = find_http_error(&chain).expect("digs through two levels");
-        assert!(
-            matches!(found, HttpError::SsrfBlocked { url } if url == "internal.corp"),
-            "typed identity preserved"
-        );
-    }
-
-    #[test]
-    fn find_http_error_none_on_foreign_chain() {
-        let chain = Wrap {
-            msg: "outer",
-            source: Box::new(std::io::Error::other("plain io")),
-        };
-        assert!(find_http_error(&chain).is_none());
-    }
-
-    #[tokio::test]
-    async fn guarded_resolver_blocks_localhost_and_serves_public() {
-        use reqwest::dns::Resolve;
-        use std::str::FromStr;
-
-        // `localhost` resolves to loopback on every OS — the resolver
-        // must refuse to hand it to the transport. (`Addrs` is not
-        // Debug — go through `.err()` instead of `expect_err`.)
-        let name = reqwest::dns::Name::from_str("localhost").expect("valid name");
-        let err = GuardedResolver
-            .resolve(name)
-            .await
-            .err()
-            .expect("localhost must be blocked by the guarded resolver");
-        let http = err
-            .downcast_ref::<HttpError>()
-            .expect("the resolver emits the kernel error type");
-        assert!(matches!(http, HttpError::SsrfBlocked { .. }), "{http}");
-    }
-
-    #[tokio::test]
-    async fn enforce_mode_blocks_loopback_end_to_end() {
-        // Through the full client: a literal loopback URL dies in the
-        // STATIC layer; `localhost` dies static too (hostname list). The
-        // resolver layer is exercised directly above — here we pin that
-        // the composed client refuses both without any network.
-        let client = ReqwestHttp::new().expect("client builds");
-        for url in ["http://127.0.0.1:9/x", "http://localhost:9/x"] {
-            let err = client
-                .get(HttpRequest::get(url))
-                .await
-                .expect_err("must be blocked before any connection");
-            assert!(matches!(err, HttpError::SsrfBlocked { .. }), "{url}: {err}");
-        }
-    }
-
-    #[tokio::test]
-    async fn disabled_mode_still_builds_a_working_client() {
-        // SsrfMode::Disabled wires the DEFAULT resolver — the build path
-        // must stay valid (connection refused ≠ build failure).
-        let mut config = HttpConfig::new();
-        config.ssrf = SsrfMode::Disabled;
-        let client = ReqwestHttp::with_config(config).expect("builds without the guard");
-        // Port 9 (discard) on loopback: NOT SsrfBlocked in Disabled mode —
-        // the failure (if any) is a plain connection error.
-        let res = client.get(HttpRequest::get("http://127.0.0.1:9/x")).await;
-        if let Err(e) = res {
-            assert!(
-                !matches!(e, HttpError::SsrfBlocked { .. }),
-                "Disabled mode must not SSRF-block: {e}"
-            );
-        }
-    }
-}
-
-/// End-to-end over REAL sockets: a minimal in-process HTTP/1.1
-/// responder (tokio `TcpListener` · zero extra deps) exercises the
-/// production client's transport mechanics — redirect loop ·
-/// cross-origin credential stripping · 303 demotion · size caps ·
-/// timeouts · streaming. `SsrfMode::Disabled` everywhere here: the
-/// targets ARE loopback (the SSRF layers have their own tests above).
-#[cfg(test)]
-mod e2e {
-    use super::*;
-    use tokio::io::{AsyncReadExt, AsyncWriteExt};
-
-    fn client(max_bytes: u64) -> ReqwestHttp {
-        let mut config = HttpConfig::new();
-        config.ssrf = SsrfMode::Disabled;
-        config.max_response_bytes = max_bytes;
-        ReqwestHttp::with_config(config).expect("client builds")
-    }
-
-    /// Serve `responses` in order on a fresh loopback listener; each
-    /// connection gets the REQUEST HEAD echoed nowhere — the canned
-    /// response verbatim. Returns the bound address.
-    async fn serve(responses: Vec<String>) -> std::net::SocketAddr {
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0")
-            .await
-            .expect("bind loopback");
-        let addr = listener.local_addr().expect("addr");
-        tokio::spawn(async move {
-            for response in responses {
-                let Ok((mut socket, _)) = listener.accept().await else {
-                    return;
-                };
-                let mut buf = [0u8; 8192];
-                let _ = socket.read(&mut buf).await;
-                let _ = socket.write_all(response.as_bytes()).await;
-                let _ = socket.shutdown().await;
-            }
-        });
-        addr
-    }
-
-    /// Serve one connection that ECHOES the received request head as
-    /// the response body (the probe for what actually went on the wire).
-    async fn serve_echo() -> std::net::SocketAddr {
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0")
-            .await
-            .expect("bind loopback");
-        let addr = listener.local_addr().expect("addr");
-        tokio::spawn(async move {
-            let Ok((mut socket, _)) = listener.accept().await else {
-                return;
-            };
-            let mut buf = vec![0u8; 8192];
-            let n = socket.read(&mut buf).await.unwrap_or(0);
-            let head = String::from_utf8_lossy(&buf[..n]).into_owned();
-            let response = format!(
-                "HTTP/1.1 200 OK\r\nContent-Length: {}\r\nConnection: close\r\n\r\n{head}",
-                head.len()
-            );
-            let _ = socket.write_all(response.as_bytes()).await;
-            let _ = socket.shutdown().await;
-        });
-        addr
-    }
-
-    fn ok_response(body: &str) -> String {
-        format!(
-            "HTTP/1.1 200 OK\r\nContent-Length: {}\r\nConnection: close\r\n\r\n{body}",
-            body.len()
-        )
-    }
-
-    #[tokio::test]
-    async fn redirect_chain_lands_with_final_url() {
-        let dest = serve(vec![ok_response("arrived")]).await;
-        let hop = serve(vec![format!(
-            "HTTP/1.1 302 Found\r\nLocation: http://{dest}/final\r\nContent-Length: 0\r\nConnection: close\r\n\r\n"
-        )])
-        .await;
-
-        let resp = client(1 << 20)
-            .get(HttpRequest::get(format!("http://{hop}/start")))
-            .await
-            .expect("redirect followed");
-        assert_eq!(resp.status, 200);
-        assert_eq!(resp.body.as_ref(), b"arrived");
-        assert_eq!(resp.final_url, format!("http://{dest}/final"));
-    }
-
-    #[tokio::test]
-    async fn cross_origin_redirect_strips_credentials_and_303_demotes() {
-        // Echo destination shows exactly what crossed the wire.
-        let dest = serve_echo().await;
-        // 303 See Other from a DIFFERENT origin (different port =
-        // different origin on loopback).
-        let hop = serve(vec![format!(
-            "HTTP/1.1 303 See Other\r\nLocation: http://{dest}/after\r\nContent-Length: 0\r\nConnection: close\r\n\r\n"
-        )])
-        .await;
-
-        let mut request = HttpRequest::post(format!("http://{hop}/submit"));
-        request
-            .headers
-            .insert("authorization".to_owned(), "Bearer sk-LEAK".to_owned());
-        request
-            .headers
-            .insert("x-trace".to_owned(), "keep-me".to_owned());
-        request.body = Some(Bytes::from_static(b"{\"payload\":1}"));
-
-        let resp = client(1 << 20).post(request).await.expect("follows");
-        let wire = String::from_utf8_lossy(&resp.body);
-        // 303 demotion: POST became GET, body gone.
-        assert!(wire.starts_with("GET /after"), "{wire}");
-        assert!(!wire.contains("payload"), "body dropped on demote: {wire}");
-        // Cross-origin: credentials stripped, neutral headers ride on.
-        assert!(!wire.contains("sk-LEAK"), "credential stripped: {wire}");
-        assert!(wire.contains("keep-me"), "neutral header kept: {wire}");
-    }
-
-    #[tokio::test]
-    async fn content_length_over_cap_fails_fast_and_counted_read_caps() {
-        // Declared Content-Length above the cap → TooLarge before the body.
-        let big = "x".repeat(64);
-        let declared = serve(vec![ok_response(&big)]).await;
-        let err = client(16)
-            .get(HttpRequest::get(format!("http://{declared}/big")))
-            .await
-            .expect_err("declared too large");
-        assert!(matches!(err, HttpError::TooLarge { max: 16, .. }), "{err}");
-
-        // No Content-Length (close-framed) → the counting reader caps.
-        let unframed = serve(vec![format!(
-            "HTTP/1.1 200 OK\r\nConnection: close\r\n\r\n{big}"
-        )])
-        .await;
-        let err = client(16)
-            .get(HttpRequest::get(format!("http://{unframed}/big")))
-            .await
-            .expect_err("counted read too large");
-        assert!(matches!(err, HttpError::TooLarge { .. }), "{err}");
-    }
-
-    /// gzip of `"hello from the compressed side"` (`mtime=0` — fully
-    /// deterministic bytes, no compressor dev-dep).
-    const GZIP_BODY: &[u8] = &[
-        0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xcb, 0x48, 0xcd, 0xc9, 0xc9,
-        0x57, 0x48, 0x2b, 0xca, 0xcf, 0x55, 0x28, 0xc9, 0x48, 0x55, 0x48, 0xce, 0xcf, 0x2d, 0x28,
-        0x4a, 0x2d, 0x2e, 0x4e, 0x4d, 0x51, 0x28, 0xce, 0x4c, 0x49, 0x05, 0x00, 0xcf, 0xe4, 0x53,
-        0xe4, 0x1e, 0x00, 0x00, 0x00,
-    ];
-
-    #[tokio::test]
-    async fn gzip_responses_decompress_transparently() {
-        // The wire carries gzip; the caller sees plain text — and the
-        // client ADVERTISED the capability (the captured request head
-        // carries Accept-Encoding).
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0")
-            .await
-            .expect("bind");
-        let addr = listener.local_addr().expect("addr");
-        let (head_tx, head_rx) = tokio::sync::oneshot::channel::<String>();
-        tokio::spawn(async move {
-            let Ok((mut socket, _)) = listener.accept().await else {
-                return;
-            };
-            let mut buf = vec![0u8; 8192];
-            let n = socket.read(&mut buf).await.unwrap_or(0);
-            let _ = head_tx.send(String::from_utf8_lossy(&buf[..n]).into_owned());
-            let head = format!(
-                "HTTP/1.1 200 OK\r\nContent-Encoding: gzip\r\nContent-Length: {}\r\nConnection: close\r\n\r\n",
-                GZIP_BODY.len()
-            );
-            let _ = socket.write_all(head.as_bytes()).await;
-            let _ = socket.write_all(GZIP_BODY).await;
-            let _ = socket.shutdown().await;
-        });
-
-        let resp = client(1 << 20)
-            .get(HttpRequest::get(format!("http://{addr}/gz")))
-            .await
-            .expect("decompresses");
-        assert_eq!(resp.body.as_ref(), b"hello from the compressed side");
-        let request_head = head_rx.await.expect("server captured the head");
-        assert!(
-            request_head
-                .to_ascii_lowercase()
-                .contains("accept-encoding"),
-            "the client advertises compression: {request_head}"
-        );
-    }
-
-    #[tokio::test]
-    async fn per_request_timeout_maps_to_timeout_error() {
-        // A listener that accepts and then stalls forever.
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0")
-            .await
-            .expect("bind");
-        let addr = listener.local_addr().expect("addr");
-        tokio::spawn(async move {
-            let Ok((socket, _)) = listener.accept().await else {
-                return;
-            };
-            tokio::time::sleep(Duration::from_secs(60)).await;
-            drop(socket);
-        });
-
-        let mut request = HttpRequest::get(format!("http://{addr}/slow"));
-        request.timeout = Some(Duration::from_millis(120));
-        let err = client(1 << 20)
-            .get(request)
-            .await
-            .expect_err("must time out");
-        assert!(matches!(err, HttpError::Timeout { .. }), "{err}");
-    }
-
-    #[tokio::test]
-    async fn streaming_delivers_chunks_and_caps_mid_stream() {
-        use futures_util_shim::collect_stream;
-
-        // Within cap: chunks arrive and concatenate.
-        let body = "streamed-body-content";
-        let addr = serve(vec![ok_response(body)]).await;
-        let stream_resp = client(1 << 20)
-            .send_streaming(HttpRequest::get(format!("http://{addr}/s")))
-            .await
-            .expect("stream opens");
-        assert_eq!(stream_resp.status, 200);
-        let collected = collect_stream(stream_resp.body).await.expect("in cap");
-        assert_eq!(collected, body.as_bytes());
-
-        // Past cap (close-framed, undeclared length): TooLarge mid-stream.
-        let big = "y".repeat(64);
-        let addr = serve(vec![format!(
-            "HTTP/1.1 200 OK\r\nConnection: close\r\n\r\n{big}"
-        )])
-        .await;
-        let stream_resp = client(16)
-            .send_streaming(HttpRequest::get(format!("http://{addr}/s")))
-            .await
-            .expect("stream opens (length unknown)");
-        let err = collect_stream(stream_resp.body)
-            .await
-            .expect_err("caps mid-stream");
-        assert!(matches!(err, HttpError::TooLarge { .. }), "{err}");
-    }
-
-    /// A tiny hand-rolled stream collector (no futures-util dev-dep):
-    /// polls the response stream to completion.
-    mod futures_util_shim {
-        use super::*;
-
-        pub(super) async fn collect_stream(
-            mut stream: Pin<Box<dyn Stream<Item = Result<Bytes, HttpError>> + Send>>,
-        ) -> Result<Vec<u8>, HttpError> {
-            let mut out = Vec::new();
-            loop {
-                let next = std::future::poll_fn(|cx| stream.as_mut().poll_next(cx)).await;
-                match next {
-                    Some(Ok(chunk)) => out.extend_from_slice(&chunk),
-                    Some(Err(e)) => return Err(e),
-                    None => return Ok(out),
-                }
-            }
-        }
-    }
-}
+mod tests;
